Updated proto-site style so we can advantage of larger screens.  It will also work for smaller screens.
for id #container1:
removed max-width:1024px; so that we take advantage of the full width of
larger screens.  The website will also scale properly with smaller
screens.
added min-width:700px;    so that lecture buttons don't wrap onto the
next line

for tag main:
removed max-width:70%; to allow main to have the full width of it's
parent container

added tag pre:
to wrap text